+++ dispatcher/src/main/kotlin/pl/droidsonroids/testing/mockwebserver/condition/PathQueryCondition.kt
@@ -10,9 +10,11 @@ import okhttp3.HttpUrl
  * @property queryParameterName query parameter name, optional
  * @property queryParameterName query parameter value for given name, optional
  */
-class PathQueryCondition(internal val path: String,
-                         internal val queryParameterName: String? = null,
-                         internal val queryParameterValue: String? = null) : HttpUrlCondition() {
+class PathQueryCondition(
+    internal val path: String,
+    internal val queryParameterName: String? = null,
+    internal val queryParameterValue: String? = null
+) : HttpUrlCondition() {
 
     override fun compareTo(other: Condition) = when {
         other is PathQueryCondition && score > other.score -> -1
@@ -23,13 +25,14 @@ class PathQueryCondition(internal val path: String,
     override fun isUrlMatching(url: HttpUrl): Boolean {
         val requestQueryParameterNames = url.queryParameterNames()
         if (url.encodedPath() == path) {
-            if (requestQueryParameterNames.contains(queryParameterName)) {
-                val requestQueryParameterValue = url.queryParameter(queryParameterName)
-                if (queryParameterValue == null || queryParameterValue == requestQueryParameterValue) {
-                    return true
+            when {
+                queryParameterName == null -> return true
+                requestQueryParameterNames.contains(queryParameterName) -> {
+                    val requestQueryParameterValue = url.queryParameter(queryParameterName)
+                    if (queryParameterValue == null || queryParameterValue == requestQueryParameterValue) {
+                        return true
+                    }
                 }
-            } else if (queryParameterName == null) {
-                return true
             }
         }
         return false