+++ CHANGELOG.md
@@ -1,7 +1,12 @@
+### 1.1.0 - 30.03.2018
+- Change `infix` in methods names to `suffix` to better reflect actual behavior
+- Change method overload to kotlin optional parameter
+- Update dependencies version
+
 ### 1.0.1 - 26.08.2017
-- Kotlin compile dependency version made explicit
-- Kotlin updated to 1.1.4-2
-- Gradle updated to 4.1
+- Specify Kotlin compile dependency version explicitly
+- Update Kotlin to 1.1.4-2
+- Update Gradle to 4.1
 
 
 ### 1.0.0 - 14.07.2017